requirements.txt
fastapi
uvicorn[standard]
anthropic
python-dotenv
python-multipart
